% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The euclidean function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric value}

\item{b}{Also a numeric value}
}
\value{
The function returns the greatest common diviser of a and b
}
\description{
Some math stuff
}
