% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{An example graph from Wikipedia.}
\format{A data frame with 18 rows and 3 variables v1, v2 and w:
\describe{
  \item{v1}{containing a vector of start nodes}
  \item{v2}{containing a vector of end nodes}
  \item{w}{containing the weight of the edges between v1 and v2}
}}
\source{
\url{https://upload.wikimedia.org/wikipedia/commons/5/57/Dijkstra_Animation.gif}
}
\usage{
wiki_graph
}
\description{
A dataset containing egde from a start node to it's neighbour nodes, from a wikipedia page (\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}) as an example of how Dijkstra's algorithm works.
}
\keyword{datasets}

