% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{The dijkstra function}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame with nodes and distances}

\item{init_node}{A numeric scalar of length 1}
}
\value{
Returns vector with distances from init_node
}
\description{
Computer science stuff (???)
}
