% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A graph that is defined by a data frame containing the three columns \code{v1}, \code{v2} and \code{w}. \code{v1} and \code{v2} are the names of the vertices in the graph and \code{w} is the weight (distance) of the path from \code{v1} to \code{v2}.}

\item{init_node}{The node from which the distances are to be calculated. \code{init_node} must be in the set of vertices defined by \code{v1} and \code{v2}.}
}
\value{
A vector containing the least weight (shortest distance) from \code{init_node} to all other vertices.
The Dijkstra's algorithm calculates the shortest distance from a given vertex in a graph and all other vertices. It is explained at the Wikipedia page \url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}.
}
\description{
Dijkstra's algorithm
}
\examples{
dijkstra(wiki_graph, 1)
# [1]  0  7  9 20 20 11
dijkstra(wiki_graph, 3)
# [1]  9 10  0 11 11  2
}
\seealso{
\code{\link{euclidean}} for our other function.
\code{\link{wiki_graph}} for the graph in the example.
}

